requirements.txt
numpy==1.18.4
portaudio>=19.6.0
pyaudio>=0.2.11
pyopengl>=3.1
